+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file. This projec
   `assertExactMetaWithoutData` assertions now assert a successful HTTP status code. Previously they were expecting
   `200 OK` though this is too restrictive for a meta-only response. However, the assertions will continue to fail for
   `204 No Content` responses because they are expecting the response to have content.
+- [#14](https://github.com/cloudcreativity/json-api-testing/issues/14) The expected `Location` header passed to the
+  `assertCreatedWithClientId` assertion can now include the expected resource id. Previously the expected header value
+  had to be passed without the id.
 
 ## [3.4.0] - 2022-01-16